Creating a dedicated mock package for the config
