+++ pkg/config/mock.go
@@ -15,73 +15,41 @@ import (
 )
 
 var (
-	isConfigMocked            = false
 	isSystemProbeConfigMocked = false
 	m                         = sync.Mutex{}
 )
 
-// MockConfig should only be used in tests
-type MockConfig struct {
+// mockConfig should only be used in tests
+type mockConfig struct {
 	Config
 }
 
 // Set is used for setting configuration in tests
-func (c *MockConfig) Set(key string, value interface{}, source model.Source) {
+func (c *mockConfig) Set(key string, value interface{}, source model.Source) {
 	c.Config.Set(key, value, source)
 }
 
 // SetWithoutSource is used for setting configuration in tests
-func (c *MockConfig) SetWithoutSource(key string, value interface{}) {
+func (c *mockConfig) SetWithoutSource(key string, value interface{}) {
 	c.Config.SetWithoutSource(key, value)
 }
 
 // SetKnown is used for setting configuration in tests
-func (c *MockConfig) SetKnown(key string) {
+func (c *mockConfig) SetKnown(key string) {
 	c.Config.SetKnown(key)
 }
 
-// Mock is creating and returning a mock config
-func Mock(t testing.TB) *MockConfig {
-	// We only check isConfigMocked when registering a cleanup function. 'isConfigMocked' avoids nested calls to
-	// Mock to reset the config to a blank state. This way we have only one mock per test and test helpers can call
-	// Mock.
-	if t != nil {
-		m.Lock()
-		defer m.Unlock()
-		if isConfigMocked {
-			// The configuration is already mocked.
-			return &MockConfig{pkgconfigsetup.Datadog()}
-		}
-
-		isConfigMocked = true
-		originalDatadogConfig := pkgconfigsetup.Datadog()
-		t.Cleanup(func() {
-			m.Lock()
-			defer m.Unlock()
-			isConfigMocked = false
-			pkgconfigsetup.SetDatadog(originalDatadogConfig)
-		})
-	}
-
-	// Configure Datadog global configuration
-	newCfg := NewConfig("datadog", "DD", strings.NewReplacer(".", "_"))
-	// Configuration defaults
-	pkgconfigsetup.SetDatadog(newCfg)
-	pkgconfigsetup.InitConfig(newCfg)
-	return &MockConfig{newCfg}
-}
-
 // MockSystemProbe is creating and returning a mock system-probe config
-func MockSystemProbe(t testing.TB) *MockConfig {
-	// We only check isConfigMocked when registering a cleanup function. 'isConfigMocked' avoids nested calls to
-	// Mock to reset the config to a blank state. This way we have only one mock per test and test helpers can call
-	// Mock.
+func MockSystemProbe(t testing.TB) model.Config {
+	// We only check isSystemProbeConfigMocked when registering a cleanup function. 'isSystemProbeConfigMocked'
+	// avoids nested calls to Mock to reset the config to a blank state. This way we have only one mock per test and
+	// test helpers can call Mock.
 	if t != nil {
 		m.Lock()
 		defer m.Unlock()
 		if isSystemProbeConfigMocked {
 			// The configuration is already mocked.
-			return &MockConfig{SystemProbe}
+			return &mockConfig{SystemProbe}
 		}
 
 		isSystemProbeConfigMocked = true
@@ -98,5 +66,5 @@ func MockSystemProbe(t testing.TB) *MockConfig {
 	SystemProbe = NewConfig("system-probe", "DD", strings.NewReplacer(".", "_"))
 	// Configuration defaults
 	pkgconfigsetup.InitSystemProbeConfig(SystemProbe)
-	return &MockConfig{SystemProbe}
+	return &mockConfig{SystemProbe}
 }
